updated queries
