+++ gatsby-node.js
@@ -105,7 +105,9 @@ exports.createSchemaCustomization = ({ actions }) => {
     type SiteSiteMetadata {
       author: Author
       siteUrl: String
-      social: Social
+      social: [Social]
+      location: String
+      role: String
     }
 
     type Author {
@@ -114,7 +116,8 @@ exports.createSchemaCustomization = ({ actions }) => {
     }
 
     type Social {
-      twitter: String
+      name: String
+      url: String
     }
 
     type MarkdownRemark implements Node {
